Finished playable version
Long overdue work in making Mission Demo playable. I was confused by this assignment, and forgot about it after I thought that it was playable. It was after working on my prototype that I notice it not being in a playable state. I almost forgot about it again with all the exams come up. Now I can say the Mission Demo is now in a playable state.